Fully transfered application's server to grpc instead of http
Added protoc file, as well as generated code to the service (this), and have rewrote it's methods in delivery layer. Adjusted ted service layer to be appropriate to usage for delivery layer. Added deleteUser functionality in repository layer. Note, that deleteUser is not yet implemented in service and delivery layers, so it is not possible to test this functionality for user (for now)